+++ packages/api/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'unchained:api',
-  version: '0.55.1',
+  version: '0.55.2',
   summary: 'Unchained Engine: GraphQL API',
   git: 'https://github.com/unchainedshop/unchained',
   documentation: 'README.md',
@@ -16,27 +16,27 @@ Package.onUse((api) => {
   api.use('ecmascript');
   api.use('webapp');
   api.use('check');
-  api.use('unchained:core@0.55.1');
-  api.use('unchained:roles@0.55.1');
-  api.use('unchained:core-currencies@0.55.1');
-  api.use('unchained:core-countries@0.55.1');
-  api.use('unchained:core-delivery@0.55.1');
-  api.use('unchained:core-discounting@0.55.1');
-  api.use('unchained:core-documents@0.55.1');
-  api.use('unchained:core-languages@0.55.1');
-  api.use('unchained:core-logger@0.55.1');
-  api.use('unchained:core-quotations@0.55.1');
-  api.use('unchained:core-orders@0.55.1');
-  api.use('unchained:core-payment@0.55.1');
-  api.use('unchained:core-pricing@0.55.1');
-  api.use('unchained:core-products@0.55.1');
-  api.use('unchained:core-accountsjs@0.55.1');
-  api.use('unchained:core-users@0.55.1');
-  api.use('unchained:core-bookmarks@0.55.1');
-  api.use('unchained:core-warehousing@0.55.1');
-  api.use('unchained:core-filters@0.55.1');
-  api.use('unchained:core-assortments@0.55.1');
-  api.use('unchained:core-subscriptions@0.55.1');
+  api.use('unchained:core@0.55.2');
+  api.use('unchained:roles@0.55.2');
+  api.use('unchained:core-currencies@0.55.2');
+  api.use('unchained:core-countries@0.55.2');
+  api.use('unchained:core-delivery@0.55.2');
+  api.use('unchained:core-discounting@0.55.2');
+  api.use('unchained:core-documents@0.55.2');
+  api.use('unchained:core-languages@0.55.2');
+  api.use('unchained:core-logger@0.55.2');
+  api.use('unchained:core-quotations@0.55.2');
+  api.use('unchained:core-orders@0.55.2');
+  api.use('unchained:core-payment@0.55.2');
+  api.use('unchained:core-pricing@0.55.2');
+  api.use('unchained:core-products@0.55.2');
+  api.use('unchained:core-accountsjs@0.55.2');
+  api.use('unchained:core-users@0.55.2');
+  api.use('unchained:core-bookmarks@0.55.2');
+  api.use('unchained:core-warehousing@0.55.2');
+  api.use('unchained:core-filters@0.55.2');
+  api.use('unchained:core-assortments@0.55.2');
+  api.use('unchained:core-subscriptions@0.55.2');
 
   api.mainModule('api.js', 'server');
 });