+++ packages/core-delivery/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'unchained:core-delivery',
-  version: '0.55.1',
+  version: '0.55.2',
   summary: 'Unchained Engine Core: Delivery',
   git: 'https://github.com/unchainedshop/unchained',
   documentation: 'README.md',
@@ -14,11 +14,11 @@ Package.onUse((api) => {
   api.use('percolate:migrations@1.0.2');
   api.use('dburles:collection-helpers@1.1.0');
   api.use('aldeed:collection2@3.2.1');
-  api.use('unchained:utils@0.55.1');
-  api.use('unchained:core-logger@0.55.1');
-  api.use('unchained:core-pricing@0.55.1');
-  api.use('unchained:core-countries@0.55.1');
-  api.use('unchained:core-worker@0.55.1');
+  api.use('unchained:utils@0.55.2');
+  api.use('unchained:core-logger@0.55.2');
+  api.use('unchained:core-pricing@0.55.2');
+  api.use('unchained:core-countries@0.55.2');
+  api.use('unchained:core-worker@0.55.2');
 
   api.mainModule('delivery.js', 'server');
 });